+++ src/unix/lwt_unix_unix.h
@@ -41,285 +41,10 @@
 #include <sys/wait.h>
 #include <unistd.h>
 
-/* +-----------------------------------------------------------------+
-   | recv/send                                                       |
-   +-----------------------------------------------------------------+ */
+#include "unix_recv_send_utils.h"
 
-static int msg_flag_table[] = {MSG_OOB, MSG_DONTROUTE, MSG_PEEK};
 
-value lwt_unix_recv(value fd, value buf, value ofs, value len, value flags)
-{
-    int ret;
-    ret = recv(Int_val(fd), &Byte(String_val(buf), Long_val(ofs)),
-               Long_val(len), caml_convert_flag_list(flags, msg_flag_table));
-    if (ret == -1) uerror("recv", Nothing);
-    return Val_int(ret);
-}
 
-value lwt_unix_bytes_recv(value fd, value buf, value ofs, value len,
-                          value flags)
-{
-    int ret;
-    ret =
-        recv(Int_val(fd), (char *)Caml_ba_array_val(buf)->data + Long_val(ofs),
-             Long_val(len), caml_convert_flag_list(flags, msg_flag_table));
-    if (ret == -1) uerror("recv", Nothing);
-    return Val_int(ret);
-}
-
-value lwt_unix_send(value fd, value buf, value ofs, value len, value flags)
-{
-    int ret;
-    ret = send(Int_val(fd), &Byte(String_val(buf), Long_val(ofs)),
-               Long_val(len), caml_convert_flag_list(flags, msg_flag_table));
-    if (ret == -1) uerror("send", Nothing);
-    return Val_int(ret);
-}
-
-value lwt_unix_bytes_send(value fd, value buf, value ofs, value len,
-                          value flags)
-{
-    int ret;
-    ret =
-        send(Int_val(fd), (char *)Caml_ba_array_val(buf)->data + Long_val(ofs),
-             Long_val(len), caml_convert_flag_list(flags, msg_flag_table));
-    if (ret == -1) uerror("send", Nothing);
-    return Val_int(ret);
-}
-
-/* +-----------------------------------------------------------------+
-   | recvfrom/sendto                                                 |
-   +-----------------------------------------------------------------+ */
-
-extern int socket_domain_table[];
-extern int socket_type_table[];
-
-value lwt_unix_recvfrom(value fd, value buf, value ofs, value len, value flags)
-{
-    CAMLparam5(fd, buf, ofs, len, flags);
-    CAMLlocal2(result, address);
-    int ret;
-    union sock_addr_union addr;
-    socklen_t addr_len;
-    addr_len = sizeof(addr);
-    ret = recvfrom(Int_val(fd), &Byte(String_val(buf), Long_val(ofs)),
-                   Long_val(len), caml_convert_flag_list(flags, msg_flag_table),
-                   &addr.s_gen, &addr_len);
-    if (ret == -1) uerror("recvfrom", Nothing);
-    address = alloc_sockaddr(&addr, addr_len, -1);
-    result = caml_alloc_tuple(2);
-    Field(result, 0) = Val_int(ret);
-    Field(result, 1) = address;
-    CAMLreturn(result);
-}
-
-value lwt_unix_bytes_recvfrom(value fd, value buf, value ofs, value len,
-                              value flags)
-{
-    CAMLparam5(fd, buf, ofs, len, flags);
-    CAMLlocal2(result, address);
-    int ret;
-    union sock_addr_union addr;
-    socklen_t addr_len;
-    addr_len = sizeof(addr);
-    ret = recvfrom(Int_val(fd), (char *)Caml_ba_data_val(buf) + Long_val(ofs),
-                   Long_val(len), caml_convert_flag_list(flags, msg_flag_table),
-                   &addr.s_gen, &addr_len);
-    if (ret == -1) uerror("recvfrom", Nothing);
-    address = alloc_sockaddr(&addr, addr_len, -1);
-    result = caml_alloc_tuple(2);
-    Field(result, 0) = Val_int(ret);
-    Field(result, 1) = address;
-    CAMLreturn(result);
-}
-
-extern void get_sockaddr(value mladdr, union sock_addr_union *addr /*out*/,
-                         socklen_t *addr_len /*out*/);
-
-value lwt_unix_sendto(value fd, value buf, value ofs, value len, value flags,
-                      value dest)
-{
-    union sock_addr_union addr;
-    socklen_t addr_len;
-    int ret;
-    get_sockaddr(dest, &addr, &addr_len);
-    ret = sendto(Int_val(fd), &Byte(String_val(buf), Long_val(ofs)),
-                 Long_val(len), caml_convert_flag_list(flags, msg_flag_table),
-                 &addr.s_gen, addr_len);
-    if (ret == -1) uerror("send", Nothing);
-    return Val_int(ret);
-}
-
-CAMLprim value lwt_unix_sendto_byte(value *argv, int argc)
-{
-    return lwt_unix_sendto(argv[0], argv[1], argv[2], argv[3], argv[4],
-                           argv[5]);
-}
-
-value lwt_unix_bytes_sendto(value fd, value buf, value ofs, value len,
-                            value flags, value dest)
-{
-    union sock_addr_union addr;
-    socklen_t addr_len;
-    int ret;
-    get_sockaddr(dest, &addr, &addr_len);
-    ret = sendto(Int_val(fd), (char *)Caml_ba_data_val(buf) + Long_val(ofs),
-                 Long_val(len), caml_convert_flag_list(flags, msg_flag_table),
-                 &addr.s_gen, addr_len);
-    if (ret == -1) uerror("send", Nothing);
-    return Val_int(ret);
-}
-
-CAMLprim value lwt_unix_bytes_sendto_byte(value *argv, int argc)
-{
-    return lwt_unix_bytes_sendto(argv[0], argv[1], argv[2], argv[3], argv[4],
-                                 argv[5]);
-}
-
-/* +-----------------------------------------------------------------+
-   | {recv/send}_msg                                                 |
-   +-----------------------------------------------------------------+ */
-
-/* Convert a caml list of io-vectors into a C array io io-vector
-   structures */
-static void store_iovs(struct iovec *iovs, value iovs_val)
-{
-    CAMLparam0();
-    CAMLlocal2(list, x);
-    for (list = iovs_val; Is_block(list); list = Field(list, 1), iovs++) {
-        x = Field(list, 0);
-        iovs->iov_base = &Byte(String_val(Field(x, 0)), Long_val(Field(x, 1)));
-        iovs->iov_len = Long_val(Field(x, 2));
-    }
-    CAMLreturn0;
-}
-
-static void bytes_store_iovs(struct iovec *iovs, value iovs_val)
-{
-    CAMLparam0();
-    CAMLlocal2(list, x);
-    for (list = iovs_val; Is_block(list); list = Field(list, 1), iovs++) {
-        x = Field(list, 0);
-        iovs->iov_base =
-            (char *)Caml_ba_data_val(Field(x, 0)) + Long_val(Field(x, 1));
-        iovs->iov_len = Long_val(Field(x, 2));
-    }
-    CAMLreturn0;
-}
-
-static value wrapper_recv_msg(int fd, int n_iovs, struct iovec *iovs)
-{
-    CAMLparam0();
-    CAMLlocal3(list, result, x);
-
-    struct msghdr msg;
-    memset(&msg, 0, sizeof(msg));
-    msg.msg_iov = iovs;
-    msg.msg_iovlen = n_iovs;
-#if defined(HAVE_FD_PASSING)
-    msg.msg_controllen = CMSG_SPACE(256 * sizeof(int));
-    msg.msg_control = alloca(msg.msg_controllen);
-    memset(msg.msg_control, 0, msg.msg_controllen);
-#endif
-
-    int ret = recvmsg(fd, &msg, 0);
-    if (ret == -1) uerror("recv_msg", Nothing);
-
-    list = Val_int(0);
-#if defined(HAVE_FD_PASSING)
-    struct cmsghdr *cm;
-    for (cm = CMSG_FIRSTHDR(&msg); cm; cm = CMSG_NXTHDR(&msg, cm))
-        if (cm->cmsg_level == SOL_SOCKET && cm->cmsg_type == SCM_RIGHTS) {
-            int *fds = (int *)CMSG_DATA(cm);
-            int nfds = (cm->cmsg_len - CMSG_LEN(0)) / sizeof(int);
-            int i;
-            for (i = nfds - 1; i >= 0; i--) {
-                x = caml_alloc_tuple(2);
-                Store_field(x, 0, Val_int(fds[i]));
-                Store_field(x, 1, list);
-                list = x;
-            };
-            break;
-        };
-#endif
-
-    result = caml_alloc_tuple(2);
-    Store_field(result, 0, Val_int(ret));
-    Store_field(result, 1, list);
-    CAMLreturn(result);
-}
-
-CAMLprim value lwt_unix_recv_msg(value val_fd, value val_n_iovs, value val_iovs)
-{
-    int n_iovs = Int_val(val_n_iovs);
-    struct iovec iovs[n_iovs];
-    store_iovs(iovs, val_iovs);
-    return wrapper_recv_msg(Int_val(val_fd), n_iovs, iovs);
-}
-
-CAMLprim value lwt_unix_bytes_recv_msg(value val_fd, value val_n_iovs,
-                                       value val_iovs)
-{
-    int n_iovs = Int_val(val_n_iovs);
-    struct iovec iovs[n_iovs];
-    bytes_store_iovs(iovs, val_iovs);
-    return wrapper_recv_msg(Int_val(val_fd), n_iovs, iovs);
-}
-
-static value wrapper_send_msg(int fd, int n_iovs, struct iovec *iovs,
-                              value val_n_fds, value val_fds)
-{
-    CAMLparam2(val_n_fds, val_fds);
-
-    struct msghdr msg;
-    memset(&msg, 0, sizeof(msg));
-    msg.msg_iov = iovs;
-    msg.msg_iovlen = n_iovs;
-
-    int n_fds = Int_val(val_n_fds);
-#if defined(HAVE_FD_PASSING)
-    if (n_fds > 0) {
-        msg.msg_controllen = CMSG_SPACE(n_fds * sizeof(int));
-        msg.msg_control = alloca(msg.msg_controllen);
-        memset(msg.msg_control, 0, msg.msg_controllen);
-
-        struct cmsghdr *cm;
-        cm = CMSG_FIRSTHDR(&msg);
-        cm->cmsg_level = SOL_SOCKET;
-        cm->cmsg_type = SCM_RIGHTS;
-        cm->cmsg_len = CMSG_LEN(n_fds * sizeof(int));
-
-        int *fds = (int *)CMSG_DATA(cm);
-        for (; Is_block(val_fds); val_fds = Field(val_fds, 1), fds++)
-            *fds = Int_val(Field(val_fds, 0));
-    };
-#else
-    if (n_fds > 0) lwt_unix_not_available("fd_passing");
-#endif
-
-    int ret = sendmsg(fd, &msg, 0);
-    if (ret == -1) uerror("send_msg", Nothing);
-    CAMLreturn(Val_int(ret));
-}
-
-CAMLprim value lwt_unix_send_msg(value val_fd, value val_n_iovs, value val_iovs,
-                                 value val_n_fds, value val_fds)
-{
-    int n_iovs = Int_val(val_n_iovs);
-    struct iovec iovs[n_iovs];
-    store_iovs(iovs, val_iovs);
-    return wrapper_send_msg(Int_val(val_fd), n_iovs, iovs, val_n_fds, val_fds);
-}
-
-CAMLprim value lwt_unix_bytes_send_msg(value val_fd, value val_n_iovs,
-                                       value val_iovs, value val_n_fds,
-                                       value val_fds)
-{
-    int n_iovs = Int_val(val_n_iovs);
-    struct iovec iovs[n_iovs];
-    bytes_store_iovs(iovs, val_iovs);
-    return wrapper_send_msg(Int_val(val_fd), n_iovs, iovs, val_n_fds, val_fds);
-}
 
 /* +-----------------------------------------------------------------+
    | Credentials                                                     |
