+++ src/unix/lwt_unix_windows.h
@@ -381,12 +381,6 @@ LWT_NOT_AVAILABLE1(unix_get_credentials)
 LWT_NOT_AVAILABLE1(unix_get_cpu)
 LWT_NOT_AVAILABLE1(unix_get_affinity)
 LWT_NOT_AVAILABLE2(unix_set_affinity)
-LWT_NOT_AVAILABLE5(unix_bytes_recv)
-LWT_NOT_AVAILABLE5(unix_bytes_send)
-LWT_NOT_AVAILABLE3(unix_bytes_recv_msg)
-LWT_NOT_AVAILABLE3(unix_bytes_send_msg)
-LWT_NOT_AVAILABLE6(unix_bytes_sendto)
-LWT_NOT_AVAILABLE6(unix_bytes_sendto_byte)
 LWT_NOT_AVAILABLE2(unix_wait_mincore_job)
 LWT_NOT_AVAILABLE1(unix_guess_blocking_job)
 LWT_NOT_AVAILABLE1(unix_open_job)
@@ -405,13 +399,6 @@ LWT_NOT_AVAILABLE1(unix_rewinddir_job)
 LWT_NOT_AVAILABLE1(unix_closedir_job)
 LWT_NOT_AVAILABLE1(unix_readlink_job)
 LWT_NOT_AVAILABLE3(unix_lockf_job)
-LWT_NOT_AVAILABLE5(unix_recv)
-LWT_NOT_AVAILABLE5(unix_send)
-LWT_NOT_AVAILABLE5(unix_recvfrom)
-LWT_NOT_AVAILABLE6(unix_sendto)
-LWT_NOT_AVAILABLE6(unix_sendto_byte)
-LWT_NOT_AVAILABLE3(unix_recv_msg)
-LWT_NOT_AVAILABLE3(unix_send_msg)
 LWT_NOT_AVAILABLE1(unix_gethostname_job)
 LWT_NOT_AVAILABLE1(unix_gethostbyname_job)
 LWT_NOT_AVAILABLE1(unix_gethostbyaddr_job)
@@ -429,7 +416,6 @@ LWT_NOT_AVAILABLE1(unix_getgrnam_job)
 LWT_NOT_AVAILABLE1(unix_getpwuid_job)
 LWT_NOT_AVAILABLE1(unix_getgrgid_job)
 LWT_NOT_AVAILABLE4(unix_wait4)
-LWT_NOT_AVAILABLE5(unix_bytes_recvfrom)
 LWT_NOT_AVAILABLE1(unix_valid_dir)
 LWT_NOT_AVAILABLE1(unix_invalidate_dir)
 LWT_NOT_AVAILABLE2(unix_bind_job)
