+++ src/unix/unix_c/unix_recv.c
@@ -1,6 +1,8 @@
 /* OCaml promise library
  * http://www.ocsigen.org/lwt
- * Copyright (C) 2010 Jérémie Dimino
+ * Copyright (C) 2009-2010 Jérémie Dimino
+ *               2009 Mauricio Fernandez
+ *               2010 Pierre Chambart
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -21,31 +23,22 @@
 
 #include "lwt_config.h"
 
-#if defined(LWT_ON_WINDOWS)
+#if !defined(LWT_ON_WINDOWS)
 
-#include "lwt_unix.h"
+#include <caml/alloc.h>
+#include <caml/mlvalues.h>
+#include <caml/unixsupport.h>
+#include <sys/types.h>
+#include <sys/socket.h>
 
-LWT_NOT_AVAILABLE1(unix_readable)
-LWT_NOT_AVAILABLE1(unix_writable)
-LWT_NOT_AVAILABLE6(unix_madvise)
-LWT_NOT_AVAILABLE4(unix_mincore)
-LWT_NOT_AVAILABLE1(unix_iov_max)
-LWT_NOT_AVAILABLE3(unix_writev)
-LWT_NOT_AVAILABLE3(unix_writev_job)
-LWT_NOT_AVAILABLE3(unix_readv)
-LWT_NOT_AVAILABLE3(unix_readv_job)
-LWT_NOT_AVAILABLE5(unix_recv)
-LWT_NOT_AVAILABLE5(unix_send)
-LWT_NOT_AVAILABLE5(unix_bytes_recv)
-LWT_NOT_AVAILABLE5(unix_bytes_send)
-LWT_NOT_AVAILABLE5(unix_recvfrom)
-LWT_NOT_AVAILABLE5(unix_bytes_recvfrom)
-LWT_NOT_AVAILABLE6(unix_sendto)
-LWT_NOT_AVAILABLE6(unix_sendto_byte)
-LWT_NOT_AVAILABLE6(unix_bytes_sendto)
-LWT_NOT_AVAILABLE6(unix_bytes_sendto_byte)
-LWT_NOT_AVAILABLE3(unix_recv_msg)
-LWT_NOT_AVAILABLE3(unix_bytes_recv_msg)
-LWT_NOT_AVAILABLE3(unix_send_msg)
-LWT_NOT_AVAILABLE3(unix_bytes_send_msg)
+#include "unix_recv_send_utils.h"
+
+value lwt_unix_recv(value fd, value buf, value ofs, value len, value flags)
+{
+    int ret;
+    ret = recv(Int_val(fd), &Byte(String_val(buf), Long_val(ofs)),
+               Long_val(len), caml_convert_flag_list(flags, msg_flag_table));
+    if (ret == -1) uerror("recv", Nothing);
+    return Val_int(ret);
+}
 #endif
